ostium-python-sdk/requirements.txt
web3>=6.0.0
gql==3.5.0
graphql-core==3.2.5
pytest  # for tests 
pytest-cov  # for coverage reports
pytest-asyncio
#
humanize
# HTTP clients for gql
aiohttp>=3.8.0
requests>=2.28.0
#
# Data handling
python-dotenv>=0.19.0  # for environment variables
pydantic>=2.0.0  # for data validation
#
# Async support
asyncio>=3.4.3

